Refactor: move shared tenant/auth types to _Shared namespaces
Reorganized tenant and auth token related types into Backend.Features.Tenants._Shared and Backend.Features.Auth._Shared namespaces. Updated all references and using statements accordingly. Changed Extensions class namespace to Krafter.Aspire.ServiceDefaults. Cleaned up redundant usings. No functional changes; improves code organization and clarity.
