+++ src/Backend/Common/Models/CurrentTenantDetails.cs
@@ -1,4 +1,4 @@
-﻿using Backend.Features.Tenants;
+using Backend.Features.Tenants._Shared;
 
 namespace Backend.Common.Models;
