+++ src/Backend/Hubs/RealtimeHub.cs
@@ -1,10 +1,10 @@
-﻿using System.Text.RegularExpressions;
+using System.Text.RegularExpressions;
 using Backend.Application.Common;
 using Backend.Common.Extensions;
 using Backend.Common.Interfaces;
 using Backend.Common.Interfaces.Auth;
 using Backend.Common.Models;
-using Backend.Features.Tenants;
+using Backend.Features.Tenants._Shared;
 using Backend.Features.Users._Shared;
 using Mapster;
 using Microsoft.AspNetCore.SignalR;
