+++ src/Backend/Infrastructure/Persistence/Tenants/TenantFinderService.cs
@@ -1,6 +1,6 @@
-﻿using Backend.Application.Common;
+using Backend.Application.Common;
 using Backend.Common.Interfaces;
-using Backend.Features.Tenants;
+using Backend.Features.Tenants._Shared;
 using Backend.Features.Users._Shared;
 using Microsoft.EntityFrameworkCore;
